+++ spec/app/models/article.rb
@@ -9,5 +9,6 @@ class Article
   attr_accessible :title, :as => [:default, :parser]
   attr_accessible :is_rss, :as => :parser
   attr_accessible :user_login
-  has_and_belongs_to_many :tags
+  has_and_belongs_to_many :tags, :validate => false
+  has_and_belongs_to_many :preferences, :inverse_of => nil, :validate => false
 end